+++ src/app.html
@@ -7,10 +7,10 @@
 		<meta name="theme-color" content="#ffffff" />
 		<link rel="manifest" href="manifest.json" />
 		<link rel="icon" type="image/png" href="icon-256.png" />
-		%svelte.head%
+		%sveltekit.head%
 	</head>
 
 	<body>
-		<div id="svelte">%svelte.body%</div>
+		<div id="svelte">%sveltekit.body%</div>
 	</body>
 </html>
