--- src/lib/Nav.svelte
@@ -1,56 +1,0 @@
-<script>
-	import { onMount } from 'svelte';
-
-	/** @type {string} */
-	let active;
-
-	onMount(() => {
-		active = window.location.pathname;
-	});
-
-	/** @param {MouseEvent} e */
-	function onClick({ target }) {
-		active = /** @type {any} */ (target).pathname;
-	}
-</script>
-
-<nav>
-	<ul on:click={onClick}>
-		<li><a class:active={active === '/'} href=".">home</a></li>
-		<li><a class:active={active === '/about'} href="about">about</a></li>
-		<li><a class:active={active === '/contact'} href="contact">contact</a></li>
-		<li><a target="blank" rel="noreferrer" href="/resume.pdf">resume</a></li>
-	</ul>
-</nav>
-
-<style>
-	nav {
-		font-weight: 400;
-	}
-
-	ul {
-		display: flex;
-		justify-content: space-between;
-		list-style: none;
-		padding: 0;
-		margin: 1em 1em;
-	}
-
-	li {
-		pointer-events: all;
-	}
-
-	.active {
-		text-decoration: none;
-	}
-
-	@media (min-width: 500px) {
-		ul {
-			justify-content: flex-start;
-			margin: 2em 2em;
-		}
-		li {
-			width: 10em;
-		}
-	}
-</style>
